Revert "Move contractgov test to tests/"
It wasn't the problem of test. backends is used for calling GovParamContract.

This reverts commit 46b12e24552099dd313b1ec59bfde3871eaf23bd.
